Updated .bash_functions file
Changed the lazygit function so if the commit message is not given as first argument,it will open the commit msg file so that user can input the MSG rather than stopping the process.
This is really crucial since most users forget the message and one should not automate the process.

If the commit message is not given,then process won't complete and the repo is not updated.
This will fix by opening that file

Resolves git push commit issue #1